Improved redis usage for websocket sessions.

+++ server/bin/shutdown-server.sh
@@ -7,13 +7,20 @@ JSBB_SERVER_PIDFILE=$JSBB_LOGDIR/server.pid
 
 if [ -e $JSBB_SERVER_PIDFILE ]; then
   JSBB_SERVER_PID=`cat $JSBB_SERVER_PIDFILE`
-  MSG=`kill -s INT $JSBB_SERVER_PID 2>&1`
-  JSBB_SERVER_SHUTDOWN_EXITCODE=$?
+  ps -p$JSBB_SERVER_PID > /dev/null
+  PS_EXIST=$?
+  if [ $PS_EXIST -eq 0 ]; then 
+    MSG=`kill -s INT $JSBB_SERVER_PID 2>&1`
+    JSBB_SERVER_SHUTDOWN_EXITCODE=$?
   
-  if [ $JSBB_SERVER_SHUTDOWN_EXITCODE -ne 0 ]; then
-    echo "Unable to shut down server. ($JSBB_SERVER_PID, $JSBB_SERVER_PIDFILE)"
-    echo "Error: $MSG"
+    if [ $JSBB_SERVER_SHUTDOWN_EXITCODE -ne 0 ]; then
+      echo "Unable to shut down server. ($JSBB_SERVER_PID, $JSBB_SERVER_PIDFILE)"
+      echo "Error: $MSG"
+    else
+      rm $JSBB_SERVER_PIDFILE
+    fi
   else
+    echo "Service is already down."
     rm $JSBB_SERVER_PIDFILE
   fi
